Creating separate component for pure list item to improve readability of sortable list components.
